move `index` to libra::internal
