+++ mercury/Cargo.toml
@@ -26,8 +26,6 @@ tokio.workspace = true
 lru-mem = "0.3.0"
 bincode = "1.3.3"
 uuid = { version = "1.7.0", features = ["v4"]}
-byteorder = "1.5.0"
-rayon =  "1.9.0"
 tracing-subscriber.workspace = true
 
 [target.'cfg(windows)'.dependencies] # only on Windows
