+++ mercury/src/internal/mod.rs
@@ -4,4 +4,3 @@
 //!
 
 pub mod pack;
-pub mod model;
